Initial refactor to DSPMain (Unfinished)
